# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "BuildX / Hippo BuildX API",
    "version": "0.2.0",
    "description": "Phase 0–2 platform + tenant admin + Property/Planning APIs.\n\n**Auth:** HttpOnly cookies or Bearer JWT.\n\n**Phase 2:** Project→Block→Tower→Floor→Unit hierarchy, bulk generation, unit status history, milestones/tasks/FS deps/Gantt, BOQ, versioned drawings/RFIs, issues, approvals, project budget.\n\n**UI guide:** docs/UI-BACKEND-GUIDE.md"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local"
    },
    {
      "url": "/",
      "description": "Current host"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "access_token",
        "description": "HttpOnly JWT set by login/session endpoints"
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/api/v1/unit-categories": {
      "get": {
        "tags": [
          "Unit Categories"
        ],
        "summary": "List unit categories",
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Unit Categories"
        ],
        "summary": "Create unit category",
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
...
  }
}